Open the MIDI handler in the background
The handler->Open() call is moved into OpenInBackground()
function that kicks off the task in a thread. This means
that we can't immediate check the result.  For that, we use
a corresponding "FinishOpening()" function that waits for
the opener thread to finish (by joining it).

The Finish() function checks if Open succeeded and
otherwise tries to open the default handler(s) using the
stand-alone OpenDefaultHandler() function.

The Finish() function is also added to the MIDI_Available()
call, which is how the MPU401 device (which accepts actual
data from the DOS softawre) queries to see if the MIDI
handler is ready to accept data.

So our handler->Open() task is kicked off during the early
init phase (before the splash screen) and is allowed to run
in parallel up until the MPU is ready and queries us.

At this point the background loaded is joined with the main
thread, at which point the original handler is ready and if
not the fallback handler will be used.  In the worst case,
even the defaults have failed and MPU considers MIDI
unavailable.