Make building of unit tests automatic or configurable
A feature is added to Meson allowing the user to explicitly
enable or disable building of unit tests:

 -Dtests=enabled
 -Dtests=disabled

If the user hasn't set this preference, then tests will
be built for debug* build types, which is the default build
type.

If the user doesn't have test-preference *and* builds a
non-debug build, such as : release, plain, or minsize, then
the unit tests will not be built.